changed location of subject files in tests because of refactoring
